v1.2-Check Mate Detection bug fixing
Fixed a bug where the isCheckMate method would register a square with an enemy piece adjacent to the king as a square where the king would not be able to move to. Changed the isCheckMate and isStalemate to properly detect squares the king could move to and squares that are in check.